Cleaned up /ui/shell/multiTabs/ file structure.
